requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
build==1.4.0
ccxt==4.5.44
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coincurve==21.0.0
cryptography==46.0.5
defirl==0.5.0
fastapi==0.118.0
fire==0.7.1
frozenlist==1.8.0
h11==0.16.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.5.3
multidict==6.7.1
nlp-project==0.0.9
numpy==1.26.4
packaging==26.0
pandas==2.3.3
propcache==0.4.1
pyaml==26.2.1
pycares==5.0.1
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.4.0
scikit-optimize==0.10.2
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
sniffio==1.3.1
starlette==0.48.0
TA-Lib==0.6.8
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.23.0
